+++ CHANGELOG.md
@@ -4,6 +4,7 @@ CHANGELOG
 1.17.0
 -------------------
 
+* IMPORTANT: PHP 7.3 or greater is now required.
 * The dependency Respect\Validation has been upgrade from 1.x to 2.1.
 
 1.16.1 (2020-11-02)